+++ Runtime/PoolModule/RootsContexts/PoolAdapterRoot.cs
@@ -1,8 +1,0 @@
-using FlowIoC.BaseModule.Root;
-
-namespace FlowIoC.PoolModule.RootsContexts
-{
-    public class PoolAdapterRoot : Root<PoolAdapterContext>
-    {
-    }
-}